@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
@@ -7,15 +7,16 @@ import { BrowserRouter } from "react-router-dom";
 import StateProvider from "./contexts/StateProvider";
 import { reducer, initialState } from "./contexts/reducer";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <StateProvider reducer={reducer} initialState={initialState}>
-        <App />
-      </StateProvider>
-    </BrowserRouter>
-  </React.StrictMode>,
-  document.getElementById("root")
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(
+  // <React.StrictMode>
+  <BrowserRouter>
+    <StateProvider reducer={reducer} initialState={initialState}>
+      <App />
+    </StateProvider>
+  </BrowserRouter>
+  // </React.StrictMode>
 );
 
 // If you want your app to work offline and load faster, you can change
